Fst {pegas}	R Documentation

F-Statistics

Description

Fst computes the Fit, Fst and Fis for each locus in the data. Rst computes the Rst for microsatellites.

Usage

Fst(x, pop = NULL, quiet = TRUE, na.alleles = "")
Rst(x, pop = NULL, quiet = TRUE, na.alleles = "")

Arguments

x
an object of class "loci".

pop
a vector or factor giving the population assignment of each row of x, or a single numeric value specifying which column of x to use as population indicator. By default, the column labelled "population" is used.

quiet
a logical value: should calculations be quiet?

na.alleles
by default, only genotypes coded as NA are considered as missing data. This option is to specify if some alleles code for missing data.

Details

Fst uses the formulae in Weir and Cockerham (1984) for each allele, and then averaged within each locus over the different alleles as suggested by these authors.

Rst uses the formulae in Slatkin (1995).

Value

A matrix with genes (loci) as rows and the three F-statistics as columns.

Author(s)

Emmanuel Paradis

References

Slatkin, M. (1995) A measure of population subdivision based on microsatellite allele frequencies. Genetics, 139, 457–462.

Weir, B. S. and Cockerham, C. C. (1984) Estimating F-statistics for the analysis of population structure. Evolution, 38, 1358–1370.

Weir, B. S. and Hill, W. G. (2002) Estimating F-statistics. Annual Review of Genetics, 36, 721–750.

See Also

fstat in package hierfstat; package dirmult on CRAN that implements various estimators of the Dirichlet-multinomial distribution, including maximum likekihood and the moments estimator of Weir and Hill (2002); Fst in Biodem that caculates Fst from a “kinship matrix”.

Examples

data(jaguar)
Fst(jaguar)
Rst(jaguar)

## no Fst but Fit and Fis in case of single population:
jaguar_corridor <- jaguar[jaguar$population == "Green Corridor", ]
Fst(jaguar_corridor)

[Package pegas version 1.4 Index]