MMD {pegas}	R Documentation

Mismatch Distribution

Description

This function draws a histogram of the frequencies of pairwise distances from a set of DNA sequences.

Usage

MMD(x, xlab = "Distance", main = "", rug = TRUE, legend = TRUE,
    lcol = c("blue", "red"), lty = c(1, 1), bw = 2, ...)

Arguments

x
a set of DNA sequences (object of class "DNAbin").

xlab
the label for the x-axis.

main
the title (none by default).

rug
a logical specifying whether to add a rug of the pairwise distances on the horizontal axis (see rug).

legend
a logical specifying whether to draw a legend.

lcol
the colours used for the curves.

lty
the line types for the curves

bw
the bandwidth used for the empirical density curve (passed to density).

...
further arguments passed to hist.

Details

The histogram shows the observed distribution of pairwise distances. The lines show an empirical density estimate (in blue) and the expected distribution under stable population (Rogers and Harpending 1992).

Value

an invisible list with three elements:

histogram
the output of the hist call.

empirical.density
the empirical density as estimated by density.

expected.curve:
the values of the curve expected under stable population.

Author(s)

Emmanuel Paradis and David Winter

References

Rogers, A. R. and Harpending, H. (1992) Population growth makes waves in the distribution of pairwise genetic-differences. Molecular Biology and Evolution, 9, 552–569.

Examples

data(woodmouse)
mmd.woodm <- MMD(woodmouse)
str(mmd.woodm)
MMD(woodmouse, breaks = 20, legend = FALSE)
MMD(woodmouse, lty = 1:2, lcol = rep("black", 2), col = "lightgrey")

[Package pegas version 1.4 Index]