R2.test {pegas}	R Documentation

Ramos-Onsins–Rozas Test of Neutrality

Description

This function computes Ramos-Onsins and Rozas's test of neutrality for a set of DNA sequences.

Usage

R2.test(x, B = 1000, theta = 1, plot = TRUE, quiet = FALSE, ...)

Arguments

x
a DNA matrix (object of class "DNAbin").

B
the number of replicates used for the simulation procedure.

theta
the value of the theta population parameter used in the simulation.

plot
a logical value specifying whether to plot the results (TRUE by default).

quiet
a logical value specifying whether to not display the progress of the simulations. The default is FALSE meaning that a progress bar is displayed by default.

...
further arguments passed to hist.

Value

a list with two elements: R2 the value of the test statistic R_2, and P.val the associated P-value. If B = 0 a single value, the test statistic, is returned

Note

The simulation procedure probably needs to be tested and improved. However the results make sense so far.

Author(s)

Emmanuel Paradis

References

Ramos-Onsins, R. and Rozas, R. (2002) Statistical properties of new neutrality tests against population growth. Molecular Biology and Evolution, 19, 2092–2100.

Sano, J. and Tachida, G. (2005) Gene genealogy and properties of test statistics of neutrality under population growth. Genetics, 169, 1687–1697.

See Also

read.dna, dist.dna

Examples

data(woodmouse)
R2.test(woodmouse, quiet = TRUE)

[Package pegas version 1.4 Index]