allelicrichness {pegas}	R Documentation

Allelic Richness and Rarefaction Plots

Description

These functions analyse allelic richness.

Usage

allelicrichness(x, pop = NULL, method = "extrapolation", min.n = NULL)
rarefactionplot(x, maxn = nrow(x), type = "l", xlab = "Sample size",
                ylab = "Expected number of alleles", plot = TRUE, ...)
rhost(x, pop = NULL, method = "extrapolation")

Arguments

x
an object of class "loci".

pop
a vector or factor giving the population assignment of each row of x, or a single numeric value specifying which column of x to use as population indicator. By default, the column labelled "population" is used.

method
a character string which should be one of “extrapolation”, “rarefaction”, “raw” or an unambiguous abbreviation of these.

min.n
the value of n used in the rarefaction method; by default, the smallest observed number of genotypes within a population.

maxn
the largest sample size used to calculate the rarefaction curve.

type, xlab, ylab
arguments passed to plot.

plot
a logical value specifying whether to do the rarefaction plot (TRUE by default).

...
arguments passed to and from methods.

Details

allelicrichness computes for each locus in x the estimated allelic richness. Three methods are available: the extrapolation method (Foulley and Ollivier 2006), the rarefaction method (Hurlbert 1971), and the raw numbers of alleles.

rarefactionplot computes the rarefaction curves of the number of alleles with respect to sample size using Hurlbert's (1971) method. A plot is made by default.

Value

allelicrichness returns a numeric matrix.

rarefactionplot returns invisibly a list of matrices with the coordinates of the rarefaction plots for each locus.

rhost returns a numeric vector.

Author(s)

Emmanuel Paradis

References

El Mousadik, A. and Petit, R. J. (1996) High level of genetic differentiation for allelic richness among populations of the argan tree [Argania spinosa (L. Skeels)] endemic to Morocco. Theoretical and Applied Genetics, 92, 832–836.

Foulley, J. L. and Ollivier, L. (2006) Estimating allelic richness and its diversity. Livestock Science, 101, 150–158.

Hurlbert, S. H. (1971) The nonconcept of species diversity: a critique and alternative parameters. Ecology, 52, 577–586.

Examples

data(jaguar)
rarefactionplot(jaguar)
allelicrichness(jaguar)
rhost(jaguar)

[Package pegas version 1.4 Index]