cophenetic.haploNet {pegas}	R Documentation

Cophenetic Matrix on Haplotype Networks

Description

This function calculates the cophenetic distance on a network. The output can be used to find nodes with short distances to most nodes.

Usage

## S3 method for class 'haploNet'
cophenetic(x)

Arguments

x
an object of class "haploNet".

Details

The results of the function are likely to be approximate in most cases with reticulations in the network. In the case of MSTs, the results are exact.

Value

a numeric matrix with colnames and rownames set to the labels of the network nodes.

Author(s)

Emmanuel Paradis

See Also

cophenetic.phylo in ape, cophenetic for the generic function

Examples

example(mst)
coph <- cophenetic(r)
rowSums(coph)

[Package pegas version 1.4 Index]