dist.hamming {pegas}	R Documentation

Hamming Distance

Description

This function implements a general purpose Hamming distance.

Usage

dist.hamming(x)

Arguments

x
a matrix or a data frame.

Details

This function should work for a wide range of data types. A typical usage would be with an object of class c("haplotype", "character").

For objects of class c("haplotype", "DNAbin"), it is better to use dist.dna(x, "n") to compute the Hamming distances.

Value

an object of class "dist".

Author(s)

Emmanuel Paradis

See Also

haplotype, dist.haplotype.loci

[Package pegas version 1.4 Index]