edit.loci {pegas}	R Documentation

Edit Allelic Data with R's Data Editor

Description

This allows to edit a data frame of class "loci" with R's spreadsheet-like data editor.

Usage

## S3 method for class 'loci'
edit(name, edit.row.names = TRUE, ...)

Arguments

name
an object of class "loci".

edit.row.names
a logical specifying to allow editing the rownames, TRUE by default (by contrast to data frames).

...
further arguments to be passed to or from other methods.

Details

This ‘method’ of the generic edit respects the class and the attribute "locicol" of the allelic data frame.

Value

A data frame with class c("loci", "data.frame").

Author(s)

Emmanuel Paradis

See Also

read.loci, summary.loci

[Package pegas version 1.4 Index]