jaguar {pegas}	R Documentation

Jaguar Micro-Satellites

Description

Fifty nine jaguars (Panthera onca) from four populations genotyped at thirteen micro-satellites by Haag et al. (2010).

Usage

data(jaguar)

Format

An object of class "loci" with 59 rows and 14 columns.

Source

Haag, T., Santos, A. S., Sana, D. A., Morato, R. G., Cullen, Jr., L., Crawshaw, Jr., P. G., De Angelo, C., Di Bitetti, M. S., Salzano, F. M. and Eizirik, E. (2010) The effect of habitat fragmentation on the genetic structure of a top predator: loss of diversity and high differentiation among remnant populations of Atlantic Forest jaguars (Panthera onca). Molecular Ecology, 22, 4906–4921.

Haag, T., Santos, A. S., Sana, D. A., Morato, R. G., Cullen, Jr., L., Crawshaw, Jr., P. G., De Angelo, C., Di Bitetti, M. S., Salzano, F. M. and Eizirik, E. (2010) Data from: The effect of habitat fragmentation on the genetic structure of a top predator: loss of diversity and high differentiation among remnant populations of Atlantic Forest jaguars (Panthera onca). Dryad Digital Repository. doi: 10.5061/dryad.1884

See Also

loci, alleles2loci

The vignette “ReadingFiles” explains how to read data like these from Dryad (https://datadryad.org/).

Examples

data(jaguar)
str(jaguar)
s <- summary(jaguar)
## Not run: 
## works if the device is large enough:
plot(s, layout = 30, las = 2)

## End(Not run)

[Package pegas version 1.4 Index]