getHaploNetOptions {pegas}	R Documentation

Options to Plot haploNet Objects

Description

These functions change the graphical options to plot haplotype networks.

Usage

getHaploNetOptions()
setHaploNetOptions(...)

Arguments

...
option(s) and value(s) to be changed (separated by commas if several).

Details

The options are listed below with their default values. Most of these values use the standard R graphical paramters (see par).

bg = "transparent": the background colour of the plot.

labels = TRUE: whether to show the haplotype labels.

labels.cex = 1: size of the haplotype labels.

labels.font = 2: font of the haplotype labels.

labels.color = "black": colour of the haplotype labels.

link.color = "black": colour of the links.

link.type = 1: type of line for the links.

link.type.alt = 2: type of lines for the alternative links.

link.width = 1: line width for the links.

link.width.alt = 1: line width for the alternative links.

haplotype.inner.color = "white": colour used inside the haplotype symbols.

haplotype.outer.color = "black": colour used for the border of the haplotype symbols.

mutations.cex = 1: size of the mutation annotations.

mutations.font = 1: font of the mutation annotations.

mutations.frame.background = "#0000FF4D": background colour (transparent blue).

mutations.frame.border = "black": colour of the frame.

mutations.text.color = 1: colour of the mutation annotations.

mutations.arrow.color = "black": colour of the arrow pointing to the link.

mutations.arrow.type = "triangle": type of the previous arrow.

mutations.sequence.color = "#BFBFBF4D": colour of the sequence (transparent grey).

mutations.sequence.end = "round": possible choices: "round", "butt", or "square" (or alternatively 0, 1, or 2).

mutations.sequence.length = 0.3: the length of the segment showing the sequence as fraction of the graphical window.

mutations.sequence.width = 5: thickness of this segment.

pie.outer.color = "black": colour of the circle around pie charts.

pie.inner.segments.color = "black": colour of the segments separating the shares of the pies.

pie.colors.function = rainbow: function used to define colours for the frequencies.

scale.ratio = 1: the scale ratio between links and symbol sizes.

show.mutation = 1: option used to show mutation or not (0).

Value

getHaploNetOptions returns a list of options. The other function returns nothing.

Author(s)

Emmanuel Paradis

See Also

plot.haploNet, mutations

Examples

getHaploNetOptions()

[Package pegas version 1.4 Index]