pegas-package {pegas}	R Documentation

Population and Evolutionary Genetics Analysis System

Description

pegas provides functions for the analysis of allelic data and of haplotype data from DNA sequences. It requires and complements two other R-packages: ape and adegenet.

The complete list of functions can be displayed with library(help = pegas).

More information on pegas can be found at https://emmanuelparadis.github.io/pegas.html.

Author(s)

Emmanuel Paradis, Thibaut Jombart, Zhian N. Kamvar, Brian Knaus, Alastair Potts, Klaus Schliep, David Winter

Maintainer: Emmanuel Paradis

[Package pegas version 1.4 Index]