plotNetMDS {pegas}	R Documentation

Plot Networks With MDS Layout

Description

This function plots a haplotype network using a layout calculated from an MDS performed on the pairwise distance matrix. The haplotypes have always the same positions for different networks.

Usage

plotNetMDS(net, d, k = 2, show.mutation = FALSE, col = NULL, font = 2, cex = 1)

Arguments

net
an object of class "haploNet".

d
an object of class "dist" (or a matrix).

k
the number of dimensions of the plot (2 or 3).

show.mutation
a logical value: if TRUE, the number of steps is printed on the links.

col
the colours of the links; by default, semi-transparent green.

font
the font used to print the labels; bold by default.

cex
the character expansion of the labels.

Value

NULL

Author(s)

Emmanuel Paradis

References

Paradis, E. (2017) Analysis of haplotype networks: the randomized minimum spanning tree method. Manuscript.

See Also

haploNet

Examples

data(woodmouse)
d <- dist.dna(woodmouse, "n")
net <- rmst(d)
plotNetMDS(net, d)

[Package pegas version 1.4 Index]