read.gtx {pegas}	R Documentation

Read Genetix Data Files

Description

This function reads allelic data from a Genetix file (.gtx).

Usage

read.gtx(file)

Arguments

file
a file name specified by either a variable of mode character or a quoted string.

Value

A data frame with class c("loci", "data.frame").

Note

The package adegenet has a similar function, read.genetix, but it returns an object of class "genind".

Author(s)

Emmanuel Paradis

References

Belkhir, K., Borsa, P., Chikhi, L., Raufaste, N. and Bonhomme, F. (1996–2004) GENETIX 4.05, logiciel sous Windows(TM) pour la genetique des populations. Laboratoire Genome, Populations, Interactions, CNRS UMR 5000, Universite de Montpellier II, Montpellier (France). https://kimura.univ-montp2.fr/genetix/

See Also

read.loci, write.loci, read.vcf, read.genetix

Examples

if (require(adegenet)) {
(X <- read.gtx(system.file("files/nancycats.gtx", package = "adegenet")))
## compare with the example in ?read.genetix
}

[Package pegas version 1.4 Index]