tajima.test {pegas}	R Documentation

Test of the Neutral Mutation Hypothesis

Description

This function tests the neutral mutation hypothesis with Tajima's D.

Usage

tajima.test(x)

Arguments

x
a set of DNA sequences (object of class "DNAbin").

Value

A list with three numeric values:

D
Tajima's D statistic.

Pval.normal
the p-value assuming that D follows a normal distribution with mean zero and variance one.

Pval.beta
the p-value assuming that D follows a beta distribution after rescaling on [0, 1] (Tajima, 1989).

Note

Alignment gaps in the sequences are ignored when calculating pairwise distances.

Author(s)

Emmanuel Paradis

References

Tajima, F. (1989) Statistical method for testing the neutral mutation hypothesis by DNA polymorphism. Genetics, 123, 595–595.

Examples

require(ape)
data(woodmouse)
tajima.test(woodmouse)

[Package pegas version 1.4 Index]