theta.msat {pegas}	R Documentation

Population Parameter THETA From Micro-Satellites

Description

This function estimates the population parameter THETA using micro-satellite data with three different estimators.

Usage

theta.msat(x)

Arguments

x
an object of class "loci".

Details

The three estimators are based on (i) the variance of the number of repeats, (ii) the expected homozygosity (both described in Kimmel et al., 1998), and (iii) the mean allele frequencies (Haasl and Payseur, 2010).

The data must be micro-satellites, so the allele names must be the allele sizes (see the example). If the data are expressed in repeat counts, then only the first estimator is affected.

Value

a numeric matrix with loci as rows and the three estimates of THETA as columns.

Author(s)

Emmanuel Paradis

References

Kimmel, M., Chakraborty, R., King, J. P., Bamshad, M., Watkins, W. S. and Jorde, L. B. (1998) Signatures of population expansion in microsatellite repeat data. Genetics, 148, 1921–1930.

Haasl, R. J. and Payseur, B. A. (2010) The number of alleles at a microsatellite defines the allele frequency spectrum and facilitates fast accurate estimation of THETA. Molecular Biology and Evolution, 27, 2702–2715.

See Also

theta.h, theta.tree

Examples

data(jaguar)
theta.msat(jaguar)

[Package pegas version 1.4 Index]