theta.s {pegas}	R Documentation

Population Parameter THETA using Segregating Sites

Description

This function computes the population parameter THETA using the number of segregating sites s in a sample of n DNA sequences.

Usage

theta.s(x, ...)
## S3 method for class 'DNAbin'
theta.s(x, variance = FALSE, ...)
## Default S3 method:
theta.s(x, n, variance = FALSE, ...)

Arguments

x
a numeric giving the number of segregating sites.

n
a numeric giving the number of sequences.

variance
a logical indicating whether the variance of the estimated THETA should be returned (TRUE), the default being FALSE.

...
arguments passed to methods.

Value

A numeric vector of length one with the estimated theta (the default), or of length two if the standard error is returned (variance = TRUE).

Author(s)

Emmanuel Paradis

References

Watterson, G. A. (1975) On the number of segragating sites in genetical models without recombination. Theoretical Population Biology, 7, 256–276.

Tajima, F. (1989) Statistical method for testing the neutral mutation hypothesis by DNA polymorphism. Genetics, 123, 585–595.

See Also

theta.h, theta.k, seg.sites, nuc.div, theta.tree

Examples

data(woodmouse)
theta.s(woodmouse)
theta.s(woodmouse, variance = TRUE)
## using the default:
s <- length(seg.sites(woodmouse))
n <- nrow(woodmouse)
theta.s(s, n)

[Package pegas version 1.4 Index]