write.loci {pegas}	R Documentation

Write Allelic Data Files

Description

This function writes allelic data into a text file.

Usage

write.loci(x, file = "", loci.sep = " ", allele.sep = "/|", ...)

Arguments

x
an object of class "loci".

file
a file name specified by either a variable of mode character, or a quoted string. By default, the data are printed on the console.

loci.sep
the character(s) use to separate the loci (columns) in the file (a space by default).

allele.sep
the character(s) used to separate the alleles for each locus in the file (a slash by default).

...
further arguments passed to write.table.

Value

NULL

Author(s)

Emmanuel Paradis

See Also

read.loci, write.table for all its options

Examples

data(jaguar)
x <- jaguar[1:10, 1:3] # take a small subset
write.loci(x)
## use of '...':
write.loci(x, loci.sep = "\t", quote = FALSE, col.names = FALSE)

[Package pegas version 1.4 Index]